DSMEngine now runs continously when started with the '-w' command line flag.  Without it DSMEngine will run only once and quit.
If an error occurs DSMEngine will return to a waiting state.

Pressing ctrl-C while it is in a waiting state no longer
terminates the application (however kill -9 still works).
(see 'TODO' comment at the end on DSMEngine:interrupt)

Commanding it to "Quit" from the web interface will kill it though.