The discWinRemote web interface.
This interface uses PHP to generate XMLRPC requests to the
dsm_server and individual DSMEngines.

I have currently tested the DSMEngine interface with a single
DSM connected to the network.

The interface first requests a list of DSMs and thier
locations from the server.  It then generates a list
of each one in a multiseelect form.

The Flight number interface at the top of the control
from is not operational yet.

DSM Start, Stop, Restart, and Quit work correctly.  You
must start DSMEngine with the '-w' flage to enable this.

Orion's and Rafsw1's Apache server currently supports PHP.

Place these files in /var/www/html/discWinRemote/.

and access them with this URL:
http://orion/discWinRemote/index.html