Added getSite() method to Variable, SampleTag.
Cache results of DSMEngine::isRTLinux.

Added NearestResampler constructor with takes a vector of const Variable*.

If a id="N" attribute is missing for a <sample> tag within a
<sensor> or <processor> then it defaults to N+1, where N
is the current number of sample tags.

StatisticsProcessor,StatisticsCruncher:  if a input variable is
specfied as a name without a site: e.g. "RH.5m" then it will
match "RH.5m" at all sites. If the site doesn't have a
station number, then the site name is appended to the output
variable name.  This allows much simpler configuration for
projects like METCRAX.