refactor scons files a bit:
Add a BUILDS option, which takes a space-separated list of architectures to
build.  Useful for limiting builds to x86 when ARM isn't needed.  Nodes are
only generated for the architectures selected in BUILDS.  The particular
architecture being built is put in the ARCH construction variable, and
src/nidas/SConscript uses that to parameterize the build and prefix paths,
rather than duplicating all that boilerplate in SConstruct with hardcorded
'arm' and 'x86' strings.

The ARM-specific compiler commands have been put in a proper scons tool,
where they might be shared and don't clutter up the SConstruct file.

Install the nidas.xsd schema into PREFIX/share/xml, where it has a chance
of being shared and referenced globally rather than being copied everywhere
it's needed.

Add options for things that were commented out, like -ggdb and XML_DEBUG.

Loop over subdirs in src/nidas rather than explicitly writing out a
SConscript call for each one.

Use PREFIX as a default for the OTHER_ prefix options, so that the OTHER
paths do not need to exist nor be specified if they are not needed.