Thread::cancel, which uses using pthread_cancel, seems to cause a segmentation fault if directed toward a thread that is in nanosleep. Substituting select() for nanosleep() works.
valgrind does not seem to be compatible with Thread::cancel() - it
hangs.  A post on the net indicates that it may have problems
with the thread cleanup handlers which are set with
pthread_cleanup_push/pop.

As a result of testing the above, left a bunch of #ifdef DEBUGs in
various places.

In DSMServer::main, put project in an auto_ptr, so it is 
automatically deleted at the end of the loop.

Tried using XmlRpcServer::exit to break out of XmlRpcServer::work.
Only works if you use a positive time arg to XmlRpcServer::work().