derive PREFIX when the arch-specific environment is created, so it's correct from the start; append the prefix paths to LIBPATH and CPPPATH in nidas/SConscript after subst(), so that only one path gets appended if the PREFIX and OTHER_PREFIX paths are the same; likewise for RPATH in the x86_env setup;
Add dynld to the LIBPATH to link to libnidas_dynld.so.  Without it, scons
will install dynld first so it can be linked into dependencies through the
install path, which also means scons will remove the installed copy with
the -c option, which is not usually what one wants from a clean.