add subdirectory iss for ISS sensors; create a TiltSensor class in ISS for the Crossbow CXTILT02 serial sensor
Split the dynld/SConscript up among the subdirectories: the objects to be
compiled and added to the dynld library just need to be added to the
LIBNIDAS_DYNLD_SOURCES construction variable of the imported env.  That way
individual subdirs have more control over specifying their source files and
how to build them.