extend the DynamicLoader and DOMObjectFactory lookup scheme to also search for symbols in shared libraries derived from the class name; see the doxygen docs for details
Only the psql library uses this capability at the moment.  All other
symbols should still be found in the libraries linked at compile-time, same
as before.  However, it's now possible to delay the loading of modules
until runtime.  For example, none of the dynld::isff modules need to be
included in libnidas_dynld or even compiled if they will not be referenced
in the configuration.