Oops, this should have been checked in with r3807.
I'll call this a "working" version w.r.t. a/d function.  It looks like
it basically keeps up with a card sampling 7 channels @ 500Hz, and doesn't put 
any significant load on the system.  However, it still seems to be very
slightly out-of-sync with the card, and leaves the card's FIFO *slowly*
filling up, reaching full after about 20-30 minutes on a quiescent 
system.