Linux driver for the ARINC card.  This was lab tested using a loop back header.
To test with the loop back add the sim_xmit entry to the xml configuration as
such:

        <arincSensor IDREF="IRS-GV" devicename="/dev/arinc0" id="3000"                    sim_xmit="true"/>
        <arincSensor IDREF="IRS-GV" devicename="/dev/arinc1" id="4000" suffix="_IRS2"     sim_xmit="true"/>

This instructs the ARINC card to generate fake labels to transmit on its two xmit channels.

I built a header plug that loops the TX to the RX via a ribbon cable.

This driver has not been tested on the aircraft yet.

NOTE - scons needs to be updated.  The makefile that builds this driver links two object files together.
For some reason the older 0.9? version of scons cannot handle this.