Re-work of building and linking libraries in scons, as suggested by Gary. When building a shareable library, specify the other libraries that it depends on. Then one just has to link executables against the shareable library and not its dependencies.
So executables can now link against -lnidas_dynld, or -lnidas, and
not have to worry about -lcap, -lpthread -ldl, etc.

As part of this rework, the nidas::util objects are only in libnidas_util.so,
and not duplicated in libnidas.so.  So DSMs will need to copy libnidas_util.so.