fix latency seen in slow sample streams:
On the sample output side, rearrange IOStream::write() so the current
buffers are immediately written if it's been more than 1/4 second since the
last write.  Likewise, on the first sample after the header has just been
written, force the sample to be written with a flush().

On the sample input side, do not block in IOStream::read() trying to fill
the buffer if there are already data in the buffer.  This is the case when
a header and the first sample have been sent together: there's no need to
block because a sample is already in the buffer.

Add tests for the IOStream::write() changes.

Update the tlogger test module so it builds and passes too.

User beware, these are fairly significant changes in the reading and
writing logic in IOStream, thus the new tests.  Run them with 'scons test'.