NOTE! The following change requires a CPLD firmware flash and a wire modification to all NCAR_A2D boards.
The JED files used for the flash are called:
VIPER_20090706.jed
VULCAN_20090706.jed

Added user ability to enable internal test voltages for the NCAR A2D board.
The user interface for this is provided by the web status page interface.

To list the controls for each A2D channel select the 'dsm_server'
Then choose 'List_NCAR_A2Ds' in the combo box below.  Press
'SUBMIT' to generate the list below.

Each channel will have a 'set' button followed by a possible voltage.
Choose a voltage then press 'set'.

To refresh the list of NCAR A2D channels press the 'SUBMIT' button
mentioned earlier.

Multiple channels can be set.  The same test voltage will be sent
to all of the set channels.

In the event that a new voltage is selected that falls outside of
the range of some of the previously set channels, those channels
will be automatically unset to prevent damage.