Added use of ioctls.  In flight the measured spectrum is not calm.
NOTE - the operator must first run "ck_lams -c" from the DSM to
to establish a baseline "calm" signature while the aircraft is
not in motion.  This signature will be subtracted from the
measured signal during the flight.

CAVEAT - this "calm" signature resides in memory.  It would
be wiped out in the event of a DSM reboot.