Added processing of the derived variable "true air speed" by the LAMS sensor.
Added a new XML config member 'TAS_level' that sets a level at which the TAS is
compaired against.  When this level is crossed the LAMS card toggles its local
oscillator.

Added (disabled) test code to the DerivedDataReader to generate a TAS value that
changes over time between two points.