Mainly a change to how a Site is associated with a DSMServer.
Previously if a DSMServer(<server>) was a direct child of Project(<project>), then all
the Sites of that project were added to that DSMServer, and a SiteIterator initialized
with the DSMServer would provide an iteration over the Sites.
This was removed, since the Sites were not really a child of the DSMServer.

A ProcessorIterator was not working correctly, it did not iterate over the Sites
if constructed from Project, and did not iterate over DSMs if constructed from a Site.

This caused a bug in the nidas::dynld::isff::PacketInputStream, which parses GOES DCP packets.

tests/iterators: a test for the iterators.

nidas/apps/data_dump: can now put -1 for dsm or sample id, and it should display
    all samples ids.