Finally, support for Bluetooth serial devices. If a DSMSerialSensor has a device name looking like:     btspp:addr[:channel] then a nidas::core::BluetootRFCommSocketIODevice will be created to do the IO.  addr can be a MAC-like 48 bit bluetooth address in colon-separated hex: xx:xx:xx:xx:xx:xx, or a "friendly" name that can be given to Bluetooth devices. In the latter case NIDAS spends several seconds listening to what devices are out there, looking for a matching friendly name.
This has just been tested on a Fedora 12 laptop at this point -
not on EL or vipers/vulcans.  More development will be needed to
support pairing of devices with PIN numbers on vipers/vulcans. On
desktops, the kbluetooth or bluetooth-applet agents handle the PIN
prompting. We'll have to implement something else on the embedded
systems, perhaps just writing the PINs to
/var/lib/bluetooth/<bdaddr>/pincodes from NIDAS or from a Debian package.