Holy const, Batman, another big checkin!
The main structural change is to get farther along in
converting Project to a non-singleton. DSMSensor,
SampleTag, Variable, VariableConverters, etc can get
to the Project by through the pointer to their parent
object. Had to use a few ugly const_cast<>s :( 
There were bugs in the initial modifications,
but hopefully the tests exposed them all.

The most numerous changes are from an effort to
reduce "#include <X.h>"s in header files if "class X;"
suffices. This should reduce the build times when a
header is modified.