More fixes in ncar_a2d driver  (will they ever end?)
At initialization time, when waiting for the inverted PPS, use
a 20 Hz IRIG callback to search for the PPS, rather than busy
waits.  Then when the second PPS is detected by the callback
schedule the FIFO read function callback.

Also, clear the FIFO between the first and second PPS, right
after switching to automatic mode, and before enabling
the sync with the PPS:

        A2DSetSYNC(brd);        // Stop A/D clocks
        A2DAuto(brd);           // Switch to automatic mode
        A2DClearFIFO(brd);      // Clear the board's FIFO...
        msleep(20);
        A2DEnable1PPS(brd);     // Enable sync with 1PPS

When using busy waits to detect the PPS, it seemed that the
first fifo read callback was getting scheduled too early.
This resulted in fifo levels that were lower than
expected before the read, and sometimes below the expected 
number of 200 words, causing the final reads of the
200 words to be from an empty fifo.

The first fifo read was perhaps being scheduled right away
at the irig interrupt that happened at the time of the PPS
detection, rather than 1/20 = 50 msec later as we thought
it was. When I added another polling delay, there should have been
400 words in the fifo, which is a level of 2 (1/4  to 1/2 full).
Instead it was still level 1, under 1/4 full (256 words).

In the old code the A2DClearFIFO was called after the busy
wait detection of the second PPS, which means that could
have been cleared late - also causing the FIFO to be less
full than expected.

This edge-of-tolerance behaviour might have differed from
boot to boot, from A2D card to A2D card.

Also disabled the reset worker. If the fifo level is
not 2 or 3, the next read by the user side returns EIO,
which will cause the device to be closed/reopened, rather
than scheduling the reset worker.

We now have a good idea of what is going on with the FIFO timing.
When I added additional polling delays, I could see 
the fifo levels grow by the levels associated with the
additional 200 words each time.