Detect measured values outside of desired level.
This is useful when running auto_cal on older cards
(i.e. the 1600 series) that may have shorted out
there internal calibration voltages during the
testing process on the 20090716 CPLD fix.

This will also expose any of those cards that are
missing a wire modifictation to its ADG417 and ADG432
chips.