Rework of the diamond dmd_mmat driver to enable generation of one or more D2A waveforms, synchronized  with the A2D sampling of one or more input channels. This simultaneous output and sampling are implemented with a "D2D" device.  Alternatively can also request the generation of a periodic waveform on an D2A, and run the A2D in its normal mode, as long as the clock rates are equivalent.
This uses much of Ryan Orendorff's additions for the generation of
the A2D waveforms. This version adds the capability to share the
cascaded clock dividers between the A2D and the D2A, using
a spinlock to prevent conflicting accesses, and enforcing that each
clock client must request the same input and output rates.

The D2D device doesn't have much code of its own, sharing
the capabilities of the A2D and the D2A, using mutexes to
avoid conflicts, and counters on the number of times
the devices are opened, in order to maintain resources.
When operating in the D2D mode, the A2D is configured as before,
except that no filtering is available.

The main change to the existing A2D driver, is that the sample circular
buffers are allocated from multiple kmallocs, rather than from one
large continuous block of memory. The number of samples in the buffers
can also be chosen as driver parameters, though I plan to add
code which should pick a good value, given the requested output
sample rate.

The ioctl commands to start/stop the A2D, D2A and D2D have been
combined to be simply DMMAT_START/DMMAT_STOP. The context of
what to start or stop is known from the destination device of the
ioctl.

This is all, as yet, untested.  That begins now...