dmd_mmat: setClock12/releaseClock12 were unbalanced and the clock could be left with a     user count > 0 after all devices were closed if a fifo overflow had occured.     Call the clock functions from the general startA2D, not the board startA2D methods.     Board startA2D functions can be void.
So far these tests have passed with a Diamond DMM-32X-AT card.

D2A output waveforms on any of the 4 channels, and any number of simultaneous output channels.
    If the user chooses 3 output channels, we have to duplicate an output.
D2A waveform output was wired back into the A2D and sampled on any of channels 1-4.
    Configured 512 point output waveform at 50 Hz, and 4 input A2D channels.
    Results look pretty good, with +- 5 counts of noise.  A2D stays in sync with D2A.

Normal A2D code works, and does setting a D2A output to a fixed value with ck_aout.

Any D2A output channel not used by waveform can be set to a fixed value while waveforms
are being output.  Can actually send a fixed value to a waveform channel too. The
waveform keeps coming. Didn't look to see if there was a spike on the scope.

There is a noticeable delay in the D2A relative to the A2D.  The last voltage of the D2A
waveform is seen as the first point on the A2D array for channel 0.  The first points on
channels 1-3 are more close to the first output voltage.  May have to set the
final output voltage in the waveform to 0.