Put some dampers on the software clock corrections.
When ndt is within the range of IRIG_MIN_DT_DIFF and IRIG_MAX_DT_DIFF, 
don't apply corrections more negative than -4 or more positive than 10.

When ndt is outside of that range, require it to be within 4 of
the ndt of the previous second before a step correction is applied.

Skipped interrupts on Vulcans generally seem to cause an ndt value of 1,2 or 3.
When greater than that, it may be due to a temporay loss of sync, so
wait to see if it is consistent.

The IRIG card on RWO at TORERO frequently looses sync, and there were several 
times in RF11, where one ndt correction was applied in one second, and
then the exact reverse of that applied in the next. This seems to be
a temporary hiccup, and it is better to not apply either correction

The limits of -4 and 10 were used on TORERO RF11.  The requirement
for ndt to be within 3 of the last when outside of
IRIG_MIN_DT_DIFF : IRIG_MAX_DT_DIFF is new.