Updates related to situation seen on dsm process at ASCII where the USB flash drive filled up on the DSM.  Writes to that FileSet then threw an IOException. The dsm process then went through the procedure of closing the output, which does an unmount of the file system, scheduling the output to be reopened in 10 seconds, which then does a mount of the ext3 file system, and proceeds to start writing samples to it, which then fails, repeat...
Somehow as this repeated, multple threads or processes (not exactly
sure which at this point) were scheduled (as seen in a ps command),
and stayed around.  After several days, one could not log into the system,
ssh reporting "resource unavailable".

A mount request starts a thread, which then does a fork/exec to
run the mount command. In general this has all worked well - but
the error scenario that has been exercised the most is the
situation of a flash drive not being present.

I have not been able to repeat this scenario in a test, which
purposefully filled up a small loop-back file system.
Extra threads/processes were not being created.  

However, in reviewing this issue, the following issues were corrected:

nidas::core::FsMount:
    Was throwing an exception a successful unmount with the libc system
    call (d'oh!). I don't think that would have caused the above issue though,
    and didn't during the small-file-system test.

In classes that implement SampleConnectionRequester::disconnect(SampleOutput*)
do separate catches of IOException when calling output->finish() and output->close(),
so that a close is done, even if finish() fails.

nidas::util::Exception,IOException:
    Moved getErrno() method from IOException to Exception base class.