Updates to support building of kernel modules for Eurotech (Arcom) Titan. User space binaries on the Titan and Viper should be identical, so the targets in $PREFIX/arm/{lib,bin} can be shared between Titans and Vipers, and they are all built once with the arm tools.
The only difference between vipers and titans should be in the kernel modules.
In order to build both titan and viper modules with arm tools, the idea of
a machine type was introduced.  The desired machine types are passed in
the command line argument MACHS, which defaults to "all", meaning all types
for an architecture.  The Environment build variable MACH is passed to
the kernel module SConscripts, with a value like "viper", "titan", "vulcan" or
"x86".

In order to avoid breaking existing boot and packaging scripts, the
viper kernel modules are installed in $PREFIX/arm/linux as before.
The Titan modules are installed in $PREFIX/arm/linux_titan.

At some point these should be changed to something like:
$PREFIX/arm/modules/viper
$PREFIX/arm/modules/titan
$PREFIX/armbe/modules/vulcan
$PREFIX/modules for x86 native.