Big rework of how Variables are associated with a Site, and how they are matched in classes like nidas::core::NearestResampler* and nidas::dynld::Statistics*.
A Variable now has a set/getSite.  A SampleTag can contain variables from
multiple Sites. 

A SampleTag will only be associated with a Site if it has a DSMConfig. Previously
the SampleTag::getSite() method tried to look everywhere for a matching station.

Added a == operator for Site, which just compares the names of the Sites.

In nidas::core::Variable, the == operator now just checks the name
of the Variable without the Site suffix, and then checks the 
equivalence of the Sites.  Before there was much buggy/confusing stuff
about station numbers.