Bug fix in the SamplePipeline. If buffering is turned off by setting rawSorterLength="0" or procSorterLength="0", and DO_BUFFERING is undefined in SampleBuffer.cc, then it was possible for the downstream SampleClient::receive() methods to be called concurrently by separate threads.
This only happened I believe if a DSMSensor::open() method does a distribute()
of one or more samples. Since the open() method is called by the
SensorOpener thread, this could be concurrent with the usual distribution
of samples of other sensors from the SensorHandler thread.

The only DSMSensor I could find that calls distribute() in its open()
method is CSAT3_Sonic.  This bug was found because that sensor is simulated
in the serial_sensor test.

I'd rather that the receive() methods not have to worry about re-entrancy.
When buffering is enabled, this is not an issue because then one thread
calls the receive() methods.

Now, when buffering is not enabled, locks are used in SampleBuffer::receive() to
enforce thread exclusion.

So now it will be part of the contract in the SamplePipeline API that multiple
threads can pass samples to it, but downstream clients don't have to worry
about re-entrancy.