Rework of some scons Configure tests.
Rather than adding CPPDEFINES to the compiler runstring as a result
of the configure tests in nidas/SConscript, the macro definitions
are now written to a header file in the build directory for each architecture,
called build*/built_headers/nidas/Config.h.

Configure autogenerates the macro names. As a result the macro names
have changed, from, for example, HAS_BZLIB_H to HAVE_BZLIB_H.

build*/built_headers is added to the CPPPATH in nidas/SConscript
when building within the nidas build tree.

nidas/Config.h gets installed to /opt/nidas/include (and to 
/opt/nidas/arm*/include for ARM).  Code that builds against nidas
and needs to check the various HAVE_ macros, should include
nidas/Config.h, with the usual CPPPATH of /opt/nidas/include.

This should be simpler than depending on a nidas scons builder
to run the Configure tests.