Do a better job of catching the situation when neither a message separator or message length are defined for a SerialSensor. SerialSensor::validate() method will throw an InvalidParameterException.
Previously this was not checked, and the message length would default to
16 bytes, just to let the user see some data. However in this situation
the Termios was left at the default, canonical mode (ICANON), which resulted
in data being lost. Now the Termios is always set to raw mode, with a length of 1.

The only time it is OK for a CharacterSensor to not have a message separator
and a length of 0 is sensor reading UDP packets. In this case it
isn't really a SerialSensor, and perhaps we should throw an exception 
related to that, but it works.