Add regex support for parsing the name of the local HCI bluetooth interface from the sensor device name.  The bluetooth socket will be bound to that interface to make the outgoing connection.  In this way we can have more than one bluetooth radio on a DSM, in order to make more than the limit of 7 connections of one radio.  If you don't bind to a specific radio the system always seems to use hci0, and returns "too many links" errno when trying to establish more than 7 connections, and does not try hci1.
Trying to connect to a bluetooth radio via its "friendly" name rather
than its 6-byte hex address may return a UnknownHostException if the
bluetooth scans have not mapped the names to addresses yet.
The open() method would throw in InvalidParameterException in this case,
rather then an IOException.  SensorHandler does not try to re-open a sensor if
an InvalidParameterException is thrown. In order that the open is retried,
cast it to an IOException instead.  Perhaps that behaviour of SensorHandler
should be re-thought, and try to re-open on an InvalidParameterException.