In callback functions which receive urbs (twod_img_rx_bulk_callback and twod_sor_rx_bulk_callback) do a wake_up_interruptible when every urb is received, rather than waiting until the output sample queue is half full, or when latencyJiffies (250 milliseconds) have elapsed since the last wakeup.
Initial tests with a spinning disk indicate this prevents the lockups seen when using a
kernel timer as the throttle function, and reduces the system slow-down when handling many
image urbs. Seem that some sort of urb starvation was happening. Needs more tests and
analysis to figure out exactly what was occurring.

DO_IRIG_TIMING pre-processor symbol is used in more places to fully support disconnecting
the pc104sg and usbtwod drivers if a kernel timer is used. Right now DO_IRIG_TIMING is
defined in usbtwod.h, but it will likely be undefined after further testing.