Cleanup of usbtwod driver, incorporating recent developments and tests.     Complete divorce of pc104sg and usbtwod drivers. usbtwod now only uses a     kernel timer to send the true air speed packets.     Added THROTTLE_JIFFIES #define and comments, to set the desired wakeup period of the     throttle function. The throttleRate is achieved by sending the correct     number of urbs in each run of the throttle function.
NOTE that some previous subversion log messagaes indicated that previous versions
of usbtwod driver would cause a crash of a Vulcan if a kernel timer callback was used
for the throttle function. That was incorrect. The throttle functions have always
been kernel timers.  The problem was when a kernel timer was used to send the true air
speed packets, at 10 Hz.  That issue has been resolved.