Support for XInclude in Xerces-c XML parser! This is parsed:
<project ...  xmlns:xi="http://www.w3.org/2001/XInclude" ...>

    <xi:include href="catalog.xml" parse="xml"/>

</project>

It is not supported in Xerces-c 2.8 which is the version on the DSMs.
However, I believe that shouldn't be a problem when the DSM is fetching
the XML from server, i.e. not in standalone mode. The server will do the
includes, and it will write the assembled XML to the DSM, which will not
have the <xi:include> elements (haven't tested this).
Will want to upgrade to Xerces-c 3 on ARM to support standalone DSMs anyway.