Added modifications for support of old DMM32AT cards, released before 32X and 32DX. They have a 512 word FIFO. The FIFO level can be queried at a resolution of half full=256 words.  As a result, the FIFO interrupt threshold must be set to >= 256. If one is scanning slowly, this will result in a delay that is likely more than the desired latency.
Changed the board type enumeration: 0=16AT,1=32AT, 2=32XAT, 3=32DXAT. Will have
to change the modprobe files.

For the DMM32, skip logic for "enhanced features" and autocal.

With the larger FIFO levels, had to increase the size of the circular buffer
of the filtered samples.

In the interrupt handler, old code was not masking the extraneous bits when checking
the interrupt status. As a result it always thought an interrupt was pending, and
went on to check the FIFO level. New code applies a mask.  Allow shared interrupts,
but the user can avoid them if they want.

Added logic to detect whether a board is present. In testing on a Titan,
reading from a PC104 address with no board present results in 0xff.
If the FPGA revision is 0xff, assume there is no board there.
Otherwise the FPGA revision is logged.