In testing of sync_server, saw some issues that looked like the result of trying to simultaneously read cal files from two threads.
The thread which is receiving the processed samples reads the
cal files in the SyncRecordSource::preLoadCalibrations() method,
in order to get information to put in the sync record header.

At the same time the thread that is producing the processed
samples is reading cal files as a routine part of processing.

Now sync_server receives the first raw sample, and calls
SyncRecordGenerator::init(), passing the time of the sample,
which then calls SyncRecordSource::preLoadCalibrations().
This should happen before processing has started.

In real-time, RawSampleService calls the virtual init() method 
on all processors, when they are started, passing
the current time. The base class SampleIOProcessor::init()
does nothing.