try to fix thread conflicts in SyncServer shutdown
When SyncServer runs as a thread in the local process, and a
SyncRecordReader instance is connected directly as a SampleClient of the
processed sample sorter, the Reader could cause a deadlock.  The pipeline
tries to flush() on SyncServer stop, but the Reader is full waiting for
samples to be read, but the reader (nimbus) has stopped reading.  In other
cases the Reader might be destroyed before the pipeline had finished
flushing, causing memory crashes.  Now the Reader checks more thoroughly if
it has been interrupted, and SyncServer makes a point of interrupting the
Reader as soon as it has been told to stop.  Likewise the processing chain
is now dismantled from end to beginning in SyncServer::stop().