flush syncGen after EOF
On input EOF in the SyncServer, the SyncRecordGenerator needs to be flushed
same as the pipeline, rather than relying on the flush to happen implicitly
when the SyncRecordGenerator is disconnected, since that happens both on
EOF and interruption, but flushes should only happen on EOF.  Also, when
SyncServer is interrupted, it needs to interrupt the rest of the processing
chain, too, including the sample pipeline and potential Reader clients, so
the chain does not deadlock while SyncServer is waiting for flushes to
finish.  This fixes the failing synctest.