Match server name when looking for filesets
Add a server name parameter to Project::findSampleOutputStreamFileSets(name)
and then return filesets corresponding to servers that exactly match
that name, or if one are found, the filesets matching the "any" server,
which has an empty name.

The method Project::findSampleOutputStreamFileSets() with no arg
does a call to uname(2) and passes the nodename to
findSampleOutputStreamFileSets(name).