Progress toward builds of Debian packages of NIDAS for x86_64 and armel
The debian directory contains initial, partly-working versions
of the source package files.

Debian loves Makefiles. I tried invoking scons from the debian/rules,
but it doesn't seem to know things like $(DESTDIR), which are
passed to the Makefile.  So the Makefile invokes scons, and
does other tasks which are done in the nidas.spec file for RPMs,
such as customizing and installing the ldconfig and pkgconfig files.

Some of the files (like etc/init.d) were moved from the rpm
directory to pkg_files so that both package systems can use them.

Updated most of the SConscript files so that header files
are installed no matter what architecture the build is for,
so that builds of nidas-dev:armel work. Maybe should just
exclude the header files from nidas-dev:armel and require
the user to install nidas-dev:amd64.

The plan is to install the binaries to the same path on both the
Debian development systems and the ARM systems.

On Debian systems:
x86_64
    /opt/nidas/bin
    /opt/nidas/include
    /opt/nidas/lib/x86_64-linux-gnu/lib*.so
    /usr/lib/x86_64-linux-gnu/pkgconfig/nidas.pc
        found automatically by pkg-config
    /lib/modules/<kernel-version>/nidas/*.ko

armel
    /opt/nidas/bin
    /opt/nidas/include
    /opt/nidas/lib/arm-linux-gnueabi
    /usr/lib/arm-linux-gnueabi/pkgconfig/nidas.pc
        probably have to set PKG_CONFIG_PATH to find it
    /lib/modules/<kernel-version>/nidas/*.ko

A "scons BUILDS=armel install" or a "apt-get install nidas:armel"
on a Debian x86_64 system will over-write the /opt/nidas/bin
directory with the armel binaries. I don't think that will
be a big issue, but might have to re-think it.
In a similar way an "apt-get install nidas-dev:armel" will
overwrite /opt/nidas/include, but they are the same as for x86_64.

On armel DSMs, the nidas-dev package won't be installed so
there should be no /opt/nidas/include or pkgconfig:
    /opt/nidas/bin
    /opt/nidas/lib/arm-linux-gnueabi
    /lib/modules/<kernel-version>/nidas/*.ko

On RedHat:
x86_64 (same as before)
    /opt/nidas/bin
    /opt/nidas/include
    /opt/nidas/lib64
    /usr/lib64/pkgconfig
old arm (same as before)
    /opt/nidas/arm/bin
    /opt/nidas/arm/lib
new armel (shouldn't ever be installed on RedHat systems)