construct LogConfig from text and get typed scheme parameters
LogConfig constructor now accepts a string which gets parsed to modify the
settings, making it more convenient to construct a LogConfig in place.
LogScheme::current() returns the current scheme and getParameterT() does
conversion of parameter values to a template type, to make it easy to
retrieve non-string parameters in code.