Reworked minor device numbering for Emerald
Previously device minor number 0 was used by two devices,
/dev/emerald0 and /dev/ttyD5, and 8 for /dev/emerald1 and ttyD13, etc.
This appears to be not compatible with device_create/device_destroy,
in that there is not a one-to-one between major/minor numbers and the
device struct.

Now, minor numbers 0-3 are for the boards, 4-11 are the 8 ports
on the first board, etc.

Setting the 232/422/485 mode is now done on the port device ttyDn
not on the board device, and so the port number is not needed
in the ioctl struct.