SampleSorter::receive() now rejects samples earlier than the sorter window.
Samples precede the current sorter window when their sample time is more
than sorter length seconds behind the the most recent sample.  Rejecting
those samples prevents the sorter output from varying depending on how
often the sorter consumer thread is scheduled.  Instead the results should
always be the same, no matter when the consumer thread ages off samples and
distributes them.  And the sorter output is guaranteed to be sorted, since
no samples can be inserted into the sample set after the sorting window has
advanced.  The discarded samples are logged same as for bogus times, and
the final discarded tally is logged with the final sorter statistics.

Because samples are actually rejected now rather than passed on out of
order, increase the default sorter lengths to 5 seconds in SyncServer to
make it a little more tolerant of time differences between sample streams.