Updates for 3.16 kernel.
Call request_mem_region instead of request_region, and do it
before ioremap.  request_region is for I/O port address, not
I/O memory and failed in 3.16. We'll see if this is backward
compatible to 2.26.

Use ioread8 instead of readb.

Assign more correct names to memory address variables.  According to
the Viper/Titan manual, addresses in the block 0x3c000000-0x3c1fffff
are physical addresses.  The result of the ioremap of the physical
address is stored in mapaddr.

Cleaned up the cleanup function. Set resources to NULL when
they have been released.