Straigtened out pc104sg IRQ confusion on Titan
Titan V2s have a bug in their CPLD register, effecting IRQS 10,11,12.
This has been fixed in the latest kernel.

The reason the status value did not have the Heartbeat bit set
was due to a bad PC104SG board, or at least a board that does not behave
like the others when there are no timing inputs.