Notify user on error, add fifoReadFrac
In testing saw situations where the driver encounted
an error, such as not being able to detect the PPS on
restart, but the user side was not notified. Added
a call to wake_up_interruptible(&brd->rwaitq_a2d) so that
the user side will know of the error.

Added a fifoReadFrac driver parameter. Testing indicates
that reading more often from the FIFO (but smaller reads)
reduces the chance of a FIFO overflow when the system is under load.
The default value is now 8 (1/8th of the FIFO). The previous
behaviour was to read about 1/4 of the FIFO on each poll.

A value of 8 results in 50 Hz polling, reading 80 words each time
A value of 4 results in 20 Hz polling reading 200 words.
(polling rate * read size) == 4000 words/second == 8 channels * 500/sec
The available polling rates from the pc104sg driver are 100/N for N integer.