Remove true air speed from input sample
True air speed is added as the first field to the CVIOutput and is not
declared as a variable in the output sample of raf.CVIProcessor.
Since TAS is read from the derived data socket, it is not a variable
from an input sensor. This removes this warning message:
    "SampleAverager: variables not found in source: CVTAS"

Reset the ostringstream to an empty string if exception occurs on the
write to the output. This fixes a memory leak if the output socket
was not connected.