Don't set module options for dmd_mmat
In the new way of doing things, under Debian 8, host specific
module options are be set in /etc/modules-load.d/*.conf. The
problem with this approach is the parameters are only passed at
boot time, not on a user modprobe.  Since that is the abbey-normal
case, the user has to know their module options when doing a
modprobe by hand, since host-specific ones are not set in the modprobe.d file.