Program to configure a Ublox NEO 5/6Q GPS via I2C
Currently one can disable or enable various NMEA messages.

This is currently a hack job, with a bunch of test
code within #ifdefs.  The basics are working:
using the SMBus ioctl to do byte-by-byte reads, and the
RDWR ioctl to do block writes. Eventually will
clean out the unused code.