Remove throw specifiers on ~Mutex, ~Cond, ~RWLock
In g++ 6, destructors default to noexcept, throw(), unless
the class contains members whose destructors throw exceptions,
in which case the destructor is declared to throw all the exceptions of
the member's destructors. This then leads to problems since a
destructor in a subclass cannot have a looser throw specifier than the base.

Since throwing exceptions in destructors is "considered harmful",
just log a message and call std::terminate(), which will cause a
stack dump.  These situations in ~Mutex, ~Cond, ~RWLock are basically
similar to referencing a bad pointer - recovery is not worth attempting.