Simplify check for board, modify EEPROM accesses
Simpliest way to check for a board is to write/read from the
APER register.

When waiting for EEPROM busy bit to clear, don't quit with -ENODEV
on a read of 0xff.  Bit 7 is the busy bit, others are not used.
Perhaps the board can get in a state where not used bits are
all 1.