Further improvements in module init function
Previous version would immediately read the config from EEPROM
and fail (-ENODEV or -ETIMEDOUT) on errors, not allowing the
module to load.

Now read the RAM config, which is initialized from EERPOM on powerup,
and check if its OK. If not OK, set it to default values, and see if those
are successfully read back. If that works, write those defaults
to EEPROM.  If the EEPROM access fails just issue warning, don't fail.

emm_read_config() is now declared void, and doesn't return -ENODEV
if it reads 0xff for a config value. These bad values will be
caught by emm_check_config().